requirement.txt
streamlit
groq
python-dotenv
feedparser
requests
beautifulsoup4
sentence-transformers
faiss-cpu
numpy
yfinance
pandas
